# swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  GenericError:
    description: GenericError is a generic error message returned by a server
    properties:
      message:
        type: string
        x-go-name: Message
    type: object
    x-go-package: go-microservice/handlers
  Team:
    description: Team defines a structure for an API team
    properties:
      city:
        description: City of the team
        type: string
        x-go-name: City
      id:
        format: int64
        type: integer
        x-go-name: ID
      name:
        description: Name of the team
        pattern: /[a-zA-Z]+/
        type: string
        x-go-name: Name
      shortName:
        description: Shorthand name of the team
        maxLength: 3
        minLength: 3
        pattern: /[A-Z]{3}/
        type: string
        x-go-name: ShortName
    required:
    - name
    - shortName
    - city
    type: object
    x-go-package: go-microservice/data
  ValidationError:
    description: ValidationError is a collection of validation error messages
    properties:
      messages:
        items:
          type: string
        type: array
        x-go-name: Messages
    type: object
    x-go-package: go-microservice/handlers
info:
  description: Documentation for Teams API
  title: Teams API
  version: 0.1.0
paths:
  /:
    get:
      description: Return a list of teams from the database
      operationId: getTeams
      responses:
        "200":
          $ref: '#/responses/teamsResponse'
      tags:
      - teams
  /{id}:
    get:
      description: Return a single team
      operationId: getTeam
      parameters:
      - description: The id of the team for which the operation relates
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/teamResponse'
        "404":
          $ref: '#/responses/errorResponse'
      tags:
      - teams
  /teams:
    post:
      description: Create a new team
      operationId: createTeam
      parameters:
      - description: |-
          team data structure to Update or Create.
          Note: the id field is ignored by update and create operations
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/Team'
      responses:
        "200":
          $ref: '#/responses/teamResponse'
        "422":
          $ref: '#/responses/errorValidation'
        "501":
          $ref: '#/responses/errorResponse'
      tags:
      - teams
    put:
      description: Update a team details
      operationId: updateTeam
      parameters:
      - description: |-
          team data structure to Update or Create.
          Note: the id field is ignored by update and create operations
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/Team'
      responses:
        "201":
          $ref: '#/responses/noContentResponse'
        "404":
          $ref: '#/responses/errorResponse'
        "422":
          $ref: '#/responses/errorValidation'
      tags:
      - teams
  /teams/{id}:
    delete:
      description: Update a teams details
      operationId: deleteTeam
      parameters:
      - description: The id of the team for which the operation relates
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "201":
          $ref: '#/responses/noContentResponse'
        "404":
          $ref: '#/responses/errorResponse'
        "501":
          $ref: '#/responses/errorResponse'
      tags:
      - teams
produces:
- application/json
responses:
  errorResponse:
    description: Generic error message returned as a string
    schema:
      $ref: '#/definitions/GenericError'
  errorValidation:
    description: Validation errors defined as an array of strings
    schema:
      $ref: '#/definitions/ValidationError'
  noContentResponse:
    description: No content is returned by this API endpoint
  teamResponse:
    description: Data structure representing a single team
    schema:
      $ref: '#/definitions/Team'
  teamsResponse:
    description: A list of teams
    schema:
      items:
        $ref: '#/definitions/Team'
      type: array
schemes:
- http
swagger: "2.0"